% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{boxplot_basic}
\alias{boxplot_basic}
\title{Basic wrapper function for boxplot}
\usage{
boxplot_basic(df, formula, labs)
}
\arguments{
\item{df}{a data.frame (or list) from which the variables
in formula should be taken.}

\item{formula}{a formula, such as y ~ grp, where y is a numeric vector of data
values to be split into groups according to the
grouping variable grp (usually a factor)}

\item{labs}{tibble object that contains the boxplot attributes}
}
\description{
Basic wrapper function for boxplot
}
