% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_packages}
\alias{check_packages}
\title{Load the require R packages and install them
if they are missing from the system}
\usage{
check_packages(requiredPackages)
}
\arguments{
\item{requirePackages}{a vector containing the require packages}
}
\description{
Load the require R packages and install them
if they are missing from the system
}
