% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_from_singaporeexpats}
\alias{data_from_singaporeexpats}
\title{Collect data from singaporeexpats website}
\usage{
data_from_singaporeexpats()
}
\value{
singaporeexpats dataset
}
\description{
Collect data from singaporeexpats website
}
