% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_from_ura}
\alias{data_from_ura}
\title{This data service will return past 3 years of median rentals of private
non-landed residential properties with at least 10 rental contracts for
the reference period.}
\usage{
data_from_ura(access_key)
}
\arguments{
\item{key}{The access key included in the email upon}
}
\value{
a tibble data frame with the json results
}
\description{
Update Frequency: End of day of every 4th Friday of January, April,
July and October. If it is a public holiday,
the data will be updated on the following working day.
}
\details{
(Variable)      Description of variable
(project)       Name of the property project
(street)        Street name that the property is on.
(x)             x coordinates of the address in SVY21 format
(y)             y coordinates of the address in SVY21 format
(rentalMedian)  Median rentals for the property
(district)      Postal district of the property
(refPeriod)     Reference period for the rental information,
                  in format of YYYYQQ
(psf25)         25th percentile rent rate per square feet per month
                  for the property for the reference period.
(median)        Median rent rate per square feet per month
                  for the property for the reference period.
(psf75)         75th percentile rent rate per square feet per month
                  for the property for the reference period.
}
