% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_condo_dataset}
\alias{get_condo_dataset}
\title{Data Collection from URA service API & singaporeexpats website}
\usage{

  get_condo_dataset(ura_access_key = "6117f3d4-81e2-4b3e-9ff9-2640045d2b5a",
  save_to = "", crawl = FALSE)
}
\arguments{
\item{ura_access_key}{access key for URA data servivce API}

\item{save_to}{path to save for the dataset}

\item{crawl}{default to F, otherwise get the dataset from actual website}
}
\value{
tibble object of the dataset
}
\description{
Data Collection from URA service API & singaporeexpats website
}
