% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_onedaytoken}
\alias{get_onedaytoken}
\title{Get an access token from URA data service api}
\usage{
get_onedaytoken(access_key)
}
\arguments{
\item{access_key}{The access key included in the email upon
successful activation of account.}
}
\value{
the token generated from the request token service
        for daily data request.
}
\description{
Get an access token from URA data service api
}
