% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_regiondistricts}
\alias{get_regiondistricts}
\title{Region to districts mapping function}
\usage{
get_regiondistricts()
}
\value{
a tibble object with region and districts mapping
}
\description{
Compute districts grouping between 3 main - regions
CCR: Core Central Region (D01,D02,D06,D09,D10,D11)
RCR: Rest of Central Region (D03,D04,D05,D07,D08,D12,D13,D14,D15,D20)
OCR: Outside Central Region (D16,D17,D18,D19,D21,D22...D28)
}
