% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ggplot_scatter}
\alias{ggplot_scatter}
\title{Scatter plot with regression lines using ggplot}
\usage{
ggplot_scatter(data, var_x, var_y, labs)
}
\arguments{
\item{var_x}{aes mapping for x axis}

\item{var_y}{aes mapping for y axis}

\item{labs}{tibble object that contains the ggplot attributes}

\item{df}{a data.frame (or list) from which the variables
in formula should be taken.}
}
\description{
Scatter plot with regression lines using ggplot
}
