% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{install_packages}
\alias{install_packages}
\title{Custom install packages function}
\usage{
install_packages(pkg)
}
\arguments{
\item{pkg}{a vector of package names to be installed}
}
\description{
Install and load multiple R packages.
Check to see if packages are installed. Install them if they are not,
then load them into the R session.
}
\examples{
install_packages (pkg)
}
