% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not_empty}
\alias{not_empty}
\title{Function to check if input path is empty}
\usage{
not_empty(file)
}
\arguments{
\item{file}{image path}
}
\value{
True if file path is specified else false
}
\description{
Function to check if input path is empty
}
