% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OLS_summary_function}
\alias{OLS_summary_function}
\title{Special Gift function from Prof Roh
For Statistical Inference using Regression}
\usage{
OLS_summary_function(my_model_estimation, mydigit)
}
\arguments{
\item{my_model_estimation}{lm model}

\item{mydigit}{round off decimal places}
}
\description{
Special Gift function from Prof Roh
For Statistical Inference using Regression
}
\examples{
OLS_summary_function (my_model_estimation, mydigit)
}
