% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_nodecontent}
\alias{parse_nodecontent}
\title{Parse node content from read_html()}
\usage{
parse_nodecontent(nodes)
}
\arguments{
\item{nodes}{xmlnode object}
}
\value{
a tibble containing condo name and TOP
}
\description{
Parse an xmlnode set object to extract out the fields
TOP and condo name
Only for content in - https://condo.singaporeexpats.com/%d/name/%s
}
