% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parse_nodecontent2}
\alias{parse_nodecontent2}
\title{Parse node content from read_html()}
\usage{
parse_nodecontent2(nodes)
}
\arguments{
\item{nodes}{xmlnode object}
}
\value{
a tibble containing condo name and TOP
}
\description{
Parse an xmlnode set object to extract out the fields
TOP and condo name
Only for content in -
https://www.singaporeexpats.com/singapore-property-pictures/photos-%s.htm
}
