% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_fit}
\alias{plot_fit}
\title{Generate 4x residual plots to test for homoscedasticity}
\usage{
plot_fit(fit, save_to = "")
}
\arguments{
\item{fit}{lm model}

\item{save_to}{path of the image to save}
}
\description{
Generate 4x residual plots to test for homoscedasticity
}
\examples{
plot_fit (fit, save_to ="")
}
