% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Generate a scatter plot with 3 regression lines using lm,
lm with poly(x,2) and loess method}
\usage{
plot_scatterplot(data, var_x, var_y, var_color = "", labs,
  regression = FALSE, save_to = "")
}
\arguments{
\item{data}{dataframe of the dataset}

\item{var_x}{independent variable}

\item{var_y}{dependent variable}

\item{var_color}{mapping color}

\item{labs}{tibble object for labs}

\item{regression}{T/F}

\item{save_to}{path of the image to save}
}
\value{
the scatterplot
}
\description{
Generate a scatter plot with 3 regression lines using lm,
lm with poly(x,2) and loess method
}
\examples{
plot_scatterplot (data, var_x, var_y, var_color="", labs, regression = F, save_to = "")
}
