% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scrape_singaporeexpats}
\alias{scrape_singaporeexpats}
\title{Perform webscraping for https://condo.singaporeexpats.com/%d/name/%s
page number
A-Z}
\usage{
scrape_singaporeexpats(url)
}
\arguments{
\item{url}{string object of the targeted url}
}
\value{
a tibble containing condo name, district and age
}
\description{
Perform webscraping for https://condo.singaporeexpats.com/%d/name/%s
page number
A-Z
}
