% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scrape_singaporeexpats2}
\alias{scrape_singaporeexpats2}
\title{Perform webscraping for
https://www.singaporeexpats.com/singapore-property-pictures/photos-%s.htm
A-G, H-S, T-Z}
\usage{
scrape_singaporeexpats2(url)
}
\arguments{
\item{url}{string object of the targeted url}
}
\value{
a tibble containing condo name, district and age
}
\description{
Perform webscraping for
https://www.singaporeexpats.com/singapore-property-pictures/photos-%s.htm
A-G, H-S, T-Z
}
